requirements.txt
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.5.2
certifi==2026.2.25
charset-normalizer==3.4.4
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.9.1
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
requests==2.32.4
schedule==1.2.2
sniffio==1.3.1
typing-extensions==4.13.2
urllib3==2.2.3
